MITO-112: integration test with ism-multisig (#12)
* assert attribute

* move to separate modules

* add evm recipient

* fix test

* testing

+++ contracts/ism-multisig/src/event.rs
@@ -12,19 +12,19 @@ pub fn emit_revoke_transfer_ownership() -> Event {
     Event::new("ism_multisig_revoke_transfer_ownership")
 }
 
-pub fn emit_enroll_validator(domain: u64, validator: String) -> Event {
+pub fn emit_enroll_validator(domain: u32, validator: String) -> Event {
     Event::new("ism_multisig_enroll_validator")
         .add_attribute("domain", domain.to_string())
         .add_attribute("validator", validator)
 }
 
-pub fn emit_unenroll_validator(domain: u64, validator: String) -> Event {
+pub fn emit_unenroll_validator(domain: u32, validator: String) -> Event {
     Event::new("ism_multisig_unenroll_validator")
         .add_attribute("domain", domain.to_string())
         .add_attribute("validator", validator)
 }
 
-pub fn emit_set_threshold(domain: u64, threshold: u8) -> Event {
+pub fn emit_set_threshold(domain: u32, threshold: u8) -> Event {
     Event::new("ism_multisig_set_threshold")
         .add_attribute("domain", domain.to_string())
         .add_attribute("threshold", threshold.to_string())
